Crate chain

Structs

ChainStore
Generic implementation of the datastore trait and structures

TipIndex
Tracks tipsets and their states by TipsetKeys and ChainEpoch.

TipSetMetadata
TipSetMetadata is the type stored as the value in the TipIndex hashmap. It contains a tipset pointing to blocks, the root cid of the chain's state after applying the messages in this tipset to it's parent state, and the cid of the receipts for these messages.

Enums

Error
Chain error

Traits

Index
Trait to allow metadata to be indexed by multiple types of structs