Struct chain::ChainStore

pub struct ChainStore<DB> {
    pub db: Arc<DB>,
    // some fields omitted
}

Stores chain data such as heaviest tipset and cached tipset info at each epoch. This structure is threadsafe, and all caches are wrapped in a mutex to allow a consistent ChainStore to be shared across tasks.

Fields
db: Arc<DB>

key-value datastore.

Implementations

impl<DB> ChainStore<DB> where
DB: BlockStore + Send + Sync + 'static,

pub fn new(db: Arc<DB>) -> Self

pub async fn set_heaviest_tipset<'_>(
&'_ self,
ts: Arc<Tipset>
) -> Result<(), Error>

Sets heaviest tipset within ChainStore and store its tipset cids under HEAD_KEY

pub async fn subscribe<'_>(&'_ self) -> (Subscriber<HeadChange>, Arc<Tipset>)

pub async fn set_tipset_tracker<'_, '_>(
&'_ self,
__arg1: &'_ BlockHeader
) -> Result<(), Error>

Sets tip_index tracker

pub fn set_genesis(&self, header: &BlockHeader) -> Result<Cid, Error>

Writes genesis to blockstore

pub async fn add_to_tipset_tracker<'_, '_>(&'_ self, header: &'_ BlockHeader)

pub async fn put_tipset<'_, '_>(&'_ self, ts: &'_ Tipset) -> Result<(), Error>

Writes tipset block headers to data store and updates heaviest tipset

pub fn genesis(&self) -> Result<Option<BlockHeader>, Error>

Returns genesis blockheader from blockstore

pub async fn heaviest_tipset<'_>(&'_ self) -> Option<Arc<Tipset>>

Returns the currently tracked heaviest tipset

pub fn publisher(&self) -> &RwLock<Publisher<HeadChange>>

pub fn blockstore(&self) -> &DB

Returns key-value store instance.

pub fn blockstore_cloned(&self) -> Arc<DB>

Clones blockstore Arc.

pub async fn tipset_from_keys<'_, '_>(
&'_ self,
tsk: &'_ TipsetKeys
) -> Result<Arc<Tipset>, Error>

Returns Tipset from key-value store from provided cids

pub fn is_block_validated(&self, cid: &Cid) -> Result<bool, Error>

Checks store if block has already been validated. Key based on the block validation prefix.

pub fn mark_block_as_validated(&self, cid: &Cid) -> Result<(), Error>

Marks block as validated in the store. This is retrieved using the block validation prefix.

pub async fn tipset_by_height<'_>(
&'_ self,
height: ChainEpoch,
ts: Arc<Tipset>,
prev: bool
) -> Result<Arc<Tipset>, Error>

Returns the tipset behind tsk at a given height. If the given height is a null round:

If prev is true, the tipset before the null round is returned.

If prev is false, the tipset following the null round is returned.

Returns None if the tipset provided was the tipset at the given height.

pub async fn get_chain_randomness<'_, '_, '_>(
&'_ self,
blocks: &'_ TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &'_ [u8]
) -> Result<[u8; 32], Box<dyn Error>>

Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch, Entropy from the ticket chain.

pub async fn get_beacon_randomness<'_, '_, '_>(
&'_ self,
blocks: &'_ TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &'_ [u8]
) -> Result<[u8; 32], Box<dyn Error>>

Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch, Entropy from the latest beacon entry.

pub async fn latest_beacon_entry<'_, '_>(
&'_ self,
ts: &'_ Tipset
) -> Result<BeaconEntry, Error>

Finds the latest beacon entry given a tipset up to 20 tipsets behind

pub fn fill_tipset(&self, ts: &Tipset) -> Option<FullTipset> where
DB: BlockStore,

Constructs and returns a full tipset if messages from storage exists - non self version

pub fn block_msgs_for_tipset(
&self,
ts: &Tipset
) -> Result<Vec<BlockMessages>, Error>

Retrieves block messages to be passed through the VM.

pub fn messages_for_tipset(
&self,
ts: &Tipset
) -> Result<Vec<ChainMessage>, Error>

Retrieves ordered valid messages from a Tipset. This will only include messages that will be passed through the VM.

pub async fn miner_load_actor_tsk<'_, '_, '_>(
&'_ self,
address: &'_ Address,
tsk: &'_ TipsetKeys
) -> Result<State, Error>

get miner state given address and tipsetkeys

pub async fn export<W, '_, '_>(
&'_ self,
tipset: &'_ Tipset,
recent_roots: ChainEpoch,
skip_old_msgs: bool,
__arg4: W
) -> Result<(), Error> where
W: AsyncWrite + Send + Unpin + 'static,

Exports a range of tipsets, as well as the state roots based on the recent_roots.