Crate chain

Modules

base_fee

headchange_json

Structs

ChainStore
Stores chain data such as heaviest tipset and cached tipset info at each epoch. This structure is threadsafe, and all caches are wrapped in a mutex to allow a consistent ChainStore to be shared across tasks.

MINIMUM_BASE_FEE
Minimum base fee amount allowed for the given Tipset.

Enums

Error
Chain error

HeadChange
Enum for pubsub channel that defines message type variant and data contained in message type.

Constants

BASE_FEE_MAX_CHANGE_DENOM
Limits gas base fee change to 12.5% of the change.

BLOCK_GAS_TARGET
Used in calculating the base fee change.

INITIAL_BASE_FEE
Genesis base fee.

PACKING_EFFICIENCY_DENOM

PACKING_EFFICIENCY_NUM

Functions

block_messages
Returns a Tuple of bls messages of type UnsignedMessage and secp messages of type SignedMessage

block_messages_from_cids
Returns a tuple of UnsignedMessage and SignedMessages from their Cid

compute_base_fee

draw_randomness
Computes a pseudorandom 32 byte Vec.

genesis
Returns the genesis block from storage.

get_chain_message
Attempts to deserialize to unsigend message or signed message and then returns it at as a

get_parent_reciept
Returns parent message receipt given block_header and message index.

messages_for_tipset
Given a tipset this function will return all unique messages in that tipset.

messages_from_cids
Returns messages from key-value store based on a slice of Cids.

persist_objects
Persists slice of serializable objects to blockstore.

read_msg_cids
Returns a tuple of cids for both Unsigned and Signed messages

set_genesis
Sets the genesis key in the BlockStore. Be careful if using this outside of the ChainStore as it will not update what the ChainStore thinks is the genesis after the ChainStore has been created.

weight
Returns the weight of provided Tipset. This function will load power actor state and calculate the total weight of the Tipset.