Struct chain::ChainStore

pub struct ChainStore<DB> {
    pub db: Arc<DB>,
    // some fields omitted
}

Stores chain data such as heaviest tipset and cached tipset info at each epoch. This structure is threadsafe, and all caches are wrapped in a mutex to allow a consistent ChainStore to be shared across tasks.

Fields
db: Arc<DB>

key-value datastore.

Implementations

impl<DB> ChainStore<DB> where
DB: BlockStore + Send + Sync + 'static,

pub fn new(db: Arc<DB>) -> Self

pub async fn set_heaviest_tipset(&self, ts: Arc<Tipset>) -> Result<(), Error>

Sets heaviest tipset within ChainStore and store its tipset cids under HEAD_KEY

pub fn set_genesis(&self, header: &BlockHeader) -> Result<Cid, Error>

Writes genesis to blockstore.

pub async fn add_to_tipset_tracker(&self, header: &BlockHeader)

Adds a BlockHeader to the tipset tracker, which tracks valid headers.

pub async fn put_tipset(&self, ts: &Tipset) -> Result<(), Error>

Writes tipset block headers to data store and updates heaviest tipset with other compatible tracked headers.

pub fn genesis(&self) -> Result<Option<BlockHeader>, Error>

Returns genesis BlockHeader from the store based on a static key.

pub async fn heaviest_tipset(&self) -> Option<Arc<Tipset>>

Returns the currently tracked heaviest tipset.

pub fn publisher(&self) -> &Publisher<HeadChange>

Returns a reference to the publisher of head changes.

pub fn blockstore(&self) -> &DB

Returns key-value store instance.

pub fn blockstore_cloned(&self) -> Arc<DB>

Clones blockstore Arc.

pub async fn tipset_from_keys(
&self,
tsk: &TipsetKeys
) -> Result<Arc<Tipset>, Error>

Returns Tipset from key-value store from provided cids

pub fn is_block_validated(&self, cid: &Cid) -> Result<bool, Error>

Checks store if block has already been validated. Key based on the block validation prefix.

pub fn mark_block_as_validated(&self, cid: &Cid) -> Result<(), Error>

Marks block as validated in the store. This is retrieved using the block validation prefix.

pub async fn tipset_by_height(
&self,
height: ChainEpoch,
ts: Arc<Tipset>,
prev: bool
) -> Result<Arc<Tipset>, Error>

Returns the tipset behind tsk at a given height. If the given height is a null round:

If prev is true, the tipset before the null round is returned.

If prev is false, the tipset following the null round is returned.

Returns None if the tipset provided was the tipset at the given height.

pub async fn get_chain_randomness_looking_forward(
&self,
blocks: &TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &[u8]
) -> Result<[u8; 32], Box<dyn Error>>

pub async fn get_chain_randomness_looking_backward(
&self,
blocks: &TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &[u8]
) -> Result<[u8; 32], Box<dyn Error>>

pub async fn get_chain_randomness(
&self,
blocks: &TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &[u8],
lookback: bool
) -> Result<[u8; 32], Box<dyn Error>>

Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch, Entropy from the ticket chain.

pub async fn get_beacon_randomness_looking_forward(
&self,
blocks: &TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &[u8]
) -> Result<[u8; 32], Box<dyn Error>>

pub async fn get_beacon_randomness_looking_backward(
&self,
blocks: &TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &[u8]
) -> Result<[u8; 32], Box<dyn Error>>

pub async fn get_beacon_randomness(
&self,
blocks: &TipsetKeys,
pers: DomainSeparationTag,
round: ChainEpoch,
entropy: &[u8],
lookback: bool
) -> Result<[u8; 32], Box<dyn Error>>

Gets 32 bytes of randomness for ChainRand paramaterized by the DomainSeparationTag, ChainEpoch, Entropy from the latest beacon entry.

pub async fn latest_beacon_entry(
&self,
ts: &Tipset
) -> Result<BeaconEntry, Error>

Finds the latest beacon entry given a tipset up to 20 tipsets behind

pub fn fill_tipset(&self, ts: &Tipset) -> Option<FullTipset> where
DB: BlockStore,

Constructs and returns a full tipset if messages from storage exists - non self version

pub fn block_msgs_for_tipset(
&self,
ts: &Tipset
) -> Result<Vec<BlockMessages>, Error>

Retrieves block messages to be passed through the VM.

pub fn messages_for_tipset(
&self,
ts: &Tipset
) -> Result<Vec<ChainMessage>, Error>

Retrieves ordered valid messages from a Tipset. This will only include messages that will be passed through the VM.

pub async fn miner_load_actor_tsk(
&self,
address: &Address,
tsk: &TipsetKeys
) -> Result<State, Error>

get miner state given address and tipsetkeys

pub async fn export<W>(
&self,
tipset: &Tipset,
recent_roots: ChainEpoch,
skip_old_msgs: bool,
writer: W
) -> Result<(), Error> where
W: AsyncWrite + Send + Unpin + 'static,

Exports a range of tipsets, as well as the state roots based on the recent_roots.

pub async fn sub_head_changes(&self) -> i64

Subscribes to head changes. This function will send the current tipset through a channel, start a task that listens to each head change event and forwards into the channel, then returns an id corresponding to the receiver of this channel from the function. This function is not blocking on events, and does not stall publishing events as it will skip over lagged events.

pub async fn next_head_change(&self, sub_id: &i64) -> Option<HeadChange>